requirements.txt
web3==6.8.0
python-dotenv==0.19.0